ai-service/requirements_full.txt
# Full requirements for real AI image recognition
# Use with Python 3.11 or 3.12 (Python 3.14 has compatibility issues)

Flask==3.1.0
flask-cors==5.0.0
Pillow==10.4.0
numpy==1.26.4
tensorflow==2.15.0
scikit-learn==1.3.2
python-dotenv==1.0.1

# Optional for enhanced features
opencv-python==4.9.0.80  # Advanced image processing
